# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /tasks:
    get:
      summary: Get all tasks
      tags:
        - tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}:
    patch:
      summary: Update an existing task by ID
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the task to update
          schema:
            type: integer
      requestBody:
        description: The task fields to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: The task was successfully updated
    delete:
      summary: Delete task
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the task to delete
          schema:
            type: integer
      responses:
        '200':
          description: The task was successfully deleted
  /users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    patch:
      summary: Update an existing user by ID
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the user to update
          schema:
            type: integer
      requestBody:
        description: The user fields to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: The user was successfully updated
    delete:
      summary: Delete user
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the user to delete
          schema:
            type: integer
      responses:
        '200':
          description: The user was successfully deleted
    get:
      summary: Get tasks by user ID
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the user
          schema:
            type: integer
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: uint
    TaskUpdate:
      type: object
      properties:
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: uint
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint
        email:
          type: string
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        password:
          type: string